feat: YjsProtocolUtil 구현
